+++ src/main/java/dev/wroud/mc/worlds/mixin/fixes/EndPortalBlockMixin.java
@@ -2,10 +2,14 @@ package dev.wroud.mc.worlds.mixin.fixes;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import dev.wroud.mc.worlds.util.DimensionDetectionUtil;
+import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.InsideBlockEffectApplier;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.EndPortalBlock;
+import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
@@ -44,6 +48,27 @@ public abstract class EndPortalBlockMixin {
     ) {
         if (DimensionDetectionUtil.isEndLikeDimension(serverLevel)) {
             return serverLevel.dimension();
+        }
+        return original;
+    }
+
+    @ModifyExpressionValue(
+            method = "entityInside",
+            at = @At(
+                    value = "FIELD",
+                    target = "Lnet/minecraft/world/level/Level;END:Lnet/minecraft/resources/ResourceKey;"
+            )
+    )
+    private ResourceKey<Level> modifyEntityInsideEndDimension(
+            ResourceKey<Level> original,
+            BlockState blockState,
+            Level level,
+            BlockPos blockPos,
+            Entity entity,
+            InsideBlockEffectApplier insideBlockEffectApplier
+    ) {
+        if (DimensionDetectionUtil.isEndLikeDimension(level)) {
+            return level.dimension();
         }
         return original;
     }